chore: extract logging and networking crates
