+++ Cargo.toml
@@ -3,6 +3,8 @@ members = [
     "safenode",
     "sn_cli",
     "sn_domain",
+    "sn_logging",
+    "sn_networking",
     "sn_protocol",
     "sn_testnet"
 ]
