+++ safenode/Cargo.toml
@@ -52,6 +52,8 @@ self_encryption = "~0.28.0"
 serde = { version = "1.0.133", features = [ "derive", "rc" ]}
 sn_dbc = { version = "17.0.0", features = ["serdes"] }
 sn_domain = { path = "../sn_domain", version = "0.1.0" }
+sn_logging = { path = "../sn_logging", version = "0.1.0" }
+sn_networking = { path = "../sn_networking", version = "0.1.0" }
 sn_protocol = { path = "../sn_protocol", version = "0.1.0" }
 thiserror = "1.0.23"
 tiny-keccak = "~2.0.2"
