+++ safenode/src/client/api.rs
@@ -11,14 +11,10 @@ use super::{
     Client, ClientEvent, ClientEventsChannel, ClientEventsReceiver, Register, RegisterOffline,
 };
 
-use crate::{
-    domain::client_transfers::SpendRequest,
-    network::multiaddr_is_global,
-    network::{close_group_majority, NetworkEvent, SwarmDriver},
-};
-
 use sn_dbc::{DbcId, SignedSpend};
 use sn_domain::client_transfers::SpendRequest;
+use sn_networking::multiaddr_is_global,
+use sn_networking::{close_group_majority, NetworkEvent, SwarmDriver};
 use sn_protocol::{
     messages::{Cmd, CmdResponse, Query, QueryResponse, Request, Response, SpendQuery},
     storage::{Chunk, ChunkAddress, DbcAddress},
