+++ safenode/src/tests_e2e/mod.rs
@@ -21,7 +21,7 @@ use eyre::Result;
 #[tokio::test(flavor = "multi_thread")]
 #[ignore = "This test is ignored because it is not stable until we have DBCs stored as records."]
 async fn multiple_sequential_transfers_succeed() -> Result<()> {
-    let _log_appender_guard = crate::log::init_node_logging(&None)?;
+    let _log_appender_guard = sn_logging::init_node_logging(&None)?;
 
     let first_wallet_dir = TempDir::new()?;
     let first_wallet_balance = Token::from_nano(1_000_000_000);
