+++ sn_cli/Cargo.toml
@@ -27,6 +27,7 @@ libp2p = { version="0.51", features = ["identify"] }
 safenode = { path = "../safenode", version = "^0.1.0" }
 sn_dbc = { version = "17.0.0", features = ["serdes"] }
 sn_domain = { path = "../sn_domain", version = "0.1.0" }
+sn_logging = { path = "../sn_logging", version = "0.1.0" }
 sn_protocol = { path = "../sn_protocol", version = "0.1.0" }
 tokio = { version = "1.17.0", features = ["fs", "io-util", "macros", "parking_lot", "rt", "sync", "time"] }
 tracing = { version = "~0.1.26" }
