+++ sn_logging/src/appender.rs
