+++ sn_logging/src/error.rs
