+++ sn_networking/src/error.rs
